+++ ChangeLog.md
@@ -3,6 +3,7 @@
 ## current
 - Bugfix: Display lifetime stats in report, not YTD
 - Force expected capitalization on player/goalie names
+- Don't show lifetime totals in report
 
 ## 0.8.0
 - Bugfix: removed quotation marks from goalie names in report